LPD-24488 Move all the logic to the service wrapper, no logic changes
